+++ evolinux-base/files/deny.sh
@@ -1,8 +1,21 @@
 #!/bin/sh
-if ! echo $1 | grep -Eq '^[0-9]+\.[0-9]+\.[0-9]+\.[0-9]+(/[0-9])?$'
+banfile=/root/ban.iptables
+if test "$#" -lt 1
 then
-	echo $1 does not look like an IPv4 address
+	printf 'usage: deny [-t] IP…\n' >&2
 	exit 1
 fi
-iptables -I INPUT -s $1 -j DROP
-echo $1 >> /root/BLACKLIST-SSH
+if test "$1" = '-t'
+then
+	banfile=/dev/null
+	shift
+fi
+for ip
+do
+	if ! echo "${ip}" | grep -Eq '^[0-9]+\.[0-9]+\.[0-9]+\.[0-9]+(/[0-9])?$'
+	then
+		printf '%s: %s does not look like an IPv4 address\n' "$0" "${ip}" >&2
+		continue
+	fi
+	printf '/sbin/iptables -I INPUT -s %s -j DROP\n' "${ip}"
+done | tee -a "${banfile}" | sh